lseek

+++ lib/user/syscall.c
@@ -101,7 +101,7 @@ int close(int fd) {
 }
 
 /* set file seek */
-int lseek(int fd, int offset, uint8_t whence) {
+off_t lseek(int fd, off_t offset, uint8_t whence) {
     return (int)_syscall3(SYS_LSEEK, fd, offset, whence);    
 }
